+++ DemoFramework/FslSimpleUI/Base/include/FslSimpleUI/Base/Gesture/GestureFlags.hpp
@@ -33,9 +33,11 @@
 
 #include <FslBase/BasicTypes.hpp>
 
+#include <cstdint>
+
 namespace Fsl::UI
 {
-  enum class GestureFlags : uint32_t
+  enum class GestureFlags : uint8_t
   {
     // A empty flag
     NotDefined = 0x00,
